feat: Add environment file to docker-compose.yml
This commit adds an environment file to the docker-compose.yml file in
order to load environment variables for the PostgreSQL container. The
environment file is specified using the `env_file` directive and points
to the `.env` file in the project's root directory. This change allows
for easier configuration of the PostgreSQL container without exposing
sensitive information in the docker-compose.yml file.